requirements.txt
flask
mastodon.py
scipy
pandas
networkx
Flask-Session
PyQt6
PyQt6-WebEngine
flask_socketio
IPython
flask_restful
flask_cors
pyinstaller
networkdisk